Change structure client and finish auth system.
