+++ client/components/header/styles.scss
@@ -55,4 +55,10 @@ header {
       border-color: $white-hover;
     }
   }
+  .userEmail {
+    float: right;
+    margin-right: 140px;
+    color: #fff;
+    margin-top: 17px;
+  }
 }
