+++ client/pages/auth/signin.js
@@ -7,9 +7,9 @@ import Button from '@material-ui/core/Button';
 import Link from "next/link";
 import Grid from '@material-ui/core/Grid';
 import {useDispatch, useSelector} from "react-redux";
-import * as types from "../../redux/types/authTypes";
-import {authValidation, resetForm, socialSignIn} from "../../redux/actions/authAction";
-import {signin} from "../../redux/actions/authAction";
+import * as types from "../../redux/types/auth/authTypes";
+import {authValidation, resetForm, socialSignIn} from "../../redux/actions/auth/authAction";
+import {signin} from "../../redux/actions/auth/authAction";
 import Alert from '@material-ui/lab/Alert';
 import FacebookLogin from 'react-facebook-login';
 import GoogleLogin from 'react-google-login';
