--- client/pages/cabinet/home.js
@@ -1,18 +1,0 @@
-import React from "react";
-import Header from "../../components/header/header";
-import Footer from "../../components/footer/footer";
-
-
-const cabinetHome = () => {
-    return (
-        <main>
-            <Header/>
-            <aside>
-                <h1>Cabinet.Home</h1>
-            </aside>
-            <Footer />
-        </main>
-    )
-}
-
-export default cabinetHome;
