+++ client/redux/reducers/auth/forgotPasswordReducer.js
@@ -1,38 +1,32 @@
-import * as types from "../types/authTypes";
+import * as typesForgotPassword from "../../types/auth/forgotPasswordTypes";
+import * as types from "../../types/auth/authTypes";
 
 const initialState = {
     email: '',
-    password: '',
     errors: [],
     message: '',
 };
 
-export const authReducer = (state = initialState, action) => {
+export const forgotPasswordReducer = (state = initialState, action) => {
     switch (action.type) {
-        case types.AUTH_CHANGE_EMAIL:
+        case typesForgotPassword.FORGOT_PASSWORD_CHANGE_EMAIL:
             return {
                 ...state,
                 email:action.email,
             }
-        case types.AUTH_CHANGE_PASSWORD:
-            return {
-                ...state,
-                password: action.password
-            }
-        case types.AUTH_VALIDATION:
+        case typesForgotPassword.FORGOT_PASSWORD_VALIDATION:
             return {
                 ...state,
                 errors: action.errors,
             }
-        case types.AUTH_SIGNUP_SUCCESS:
+        case typesForgotPassword.FORGOT_PASSWORD_SUCCESS:
             return {
                 ...state,
-                message: action.message,
+                message:action.message,
             }
-        case types.AUTH_RESET_FORM:
+        case typesForgotPassword.FORGOT_PASSWORD_RESET_FORM:
             return {
                 email: '',
-                password: '',
                 message: '',
                 errors: [],
             }
