+++ client/redux/reducers/auth/tokenReducer.js
@@ -1,4 +1,4 @@
-import * as types from "../types/tokenTypes";
+import * as types from "../../types/auth/tokenTypes";
 
 const initialState = {
     token: '',
@@ -10,6 +10,11 @@ export const tokenReducer = (state = initialState, action) => {
             return {
                 ...state,
                 token:action.token,
+            }
+        case types.TOKEN_REMOVE_TOKEN:
+            return {
+                ...state,
+                token:'',
             }
         default:
             return state;
