+++ client/redux/reducers/index.js
@@ -1,10 +1,10 @@
 import {combineReducers} from "redux";
-import {authReducer} from "./authReducer";
-import {tokenReducer} from "./tokenReducer";
-import {confirmRegisterReducer} from "./confirmRegisterReducer";
-import {forgotPasswordReducer} from "./forgotPasswordReducer";
-import {newPasswordReducer} from "./newPasswordReducer";
-import {userInfoReducer} from "./userInfoReducer";
+import {authReducer} from "./auth/authReducer";
+import {tokenReducer} from "./auth/tokenReducer";
+import {confirmRegisterReducer} from "./auth/confirmRegisterReducer";
+import {forgotPasswordReducer} from "./auth/forgotPasswordReducer";
+import {newPasswordReducer} from "./auth/newPasswordReducer";
+import {userInfoReducer} from "./auth/userInfoReducer";
 
 export default combineReducers({
     auth: authReducer,
