+++ client/redux/types/auth/confirmRegisterTypes.js
