+++ client/redux/types/auth/forgotPasswordTypes.js
