+++ client/redux/types/auth/tokenTypes.js
@@ -1,0 +1,2 @@
+export const TOKEN_SET_TOKEN = 'TOKEN_SET_TOKEN';
+export const TOKEN_REMOVE_TOKEN = 'TOKEN_REMOVE_TOKEN';
