+++ client/redux/types/auth/userInfoTypes.js
@@ -1,0 +1,2 @@
+export const USER_INFO_SET_DATA = 'USER_INFO_SET_DATA';
+export const USER_INFO_REMOVE_DATA = 'USER_INFO_REMOVE_DATA';
