--- client/redux/types/userInfoTypes.js
@@ -1,1 +1,0 @@
-export const USER_INFO_SET_DATA = 'USER_INFO_SET_DATA';
